การเพิ่ม location จะเพิ่ม stock สำหรับ location เป็น default ให้ด้วยที่ตาราง stocks.

+++ app/Location.php
@@ -7,35 +7,59 @@ use Illuminate\Database\Eloquent\Model;
 class Location extends Model
 {
     protected $fillable = [
-    	'name',
+        'name',
     ];
+
+    public static function boot()
+    {
+        parent::boot();
+
+        static::created(function ($model) {
+            $products = Product::all();
+            foreach ($products as $product) {
+                $model->stocks()->create([
+                    'product_id' => $product->id,
+                    'qty'        => 0,
+                ]);
+            }
+        });
+    }
 
     public function receiveItems()
     {
-    	return $this->hasMany(ReceiveItem::class, 'location_id', 'id');
+        return $this->hasMany(ReceiveItem::class, 'location_id', 'id');
+    }
+
+    public function stocks()
+    {
+        return $this->hasMany(Stock::class);
     }
 
     public static function deleteByCondition($id)
     {
-    	$query = self::with([
-    			'receiveItems',
-    		])
-    		->where('id', $id)
-    		->first();
+        $query = self::with([
+            'receiveItems',
+            'stocks' => function ($query) {
+                $query->where('qty', '>', 0);
+            },
+        ])
+            ->where('id', $id)
+            ->first();
 
-    	if($query->receiveItems()->count())
-			return [
-				'status' => false,
-				'title' => trans('location.label.name'),
-				'message' => trans('location.message_alert.delete_unsuccess'),
-			];
+        if ($query->receiveItems->count() || $query->stocks->count()) {
+            return [
+                'status'  => false,
+                'title'   => trans('location.label.name'),
+                'message' => trans('location.message_alert.delete_unsuccess'),
+            ];
+        }
 
-		$response = $query->delete();
+        $response = $query->delete();
 
-		return [
-			'status' => $response,
-			'title' => trans('location.label.name'),
-			'message' => trans('location.message_alert.delete_success'),
-		];
+        return [
+            'status'  => $response,
+            'title'   => trans('location.label.name'),
+            'message' => trans('location.message_alert.delete_success'),
+        ];
     }
 }
